feat: home header with new children welcome
