+++ feature/dashboard/src/main/java/com/odisby/feature/dashboard/ui/DashboardScreen.kt
@@ -109,277 +109,3 @@ fun DashboardScreen(
         }
     }
 }
-
-@Composable
-fun AppBar(childFirstName: String, onMenuClick: () -> Unit) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 12.dp, start = 26.dp, bottom = 32.dp, end = 26.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.Top,
-    ) {
-        Row {
-            Image(
-                painter = painterResource(id = R.drawable.logo_home),
-                contentDescription = stringResource(id = R.string.logo_home),
-                contentScale = ContentScale.Fit,
-                modifier = Modifier.size(62.dp)
-            )
-            Spacer(modifier = Modifier.width(12.dp))
-            Text(
-                fontSize = 24.sp,
-                text = stringResource(id = R.string.hello, childFirstName)
-            )
-        }
-        MenuIcon(onClick = onMenuClick)
-    }
-}
-
-@Composable
-fun MenuIcon(onClick: () -> Unit) {
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier = Modifier
-            .size(32.dp)
-            .clip(RoundedCornerShape(20))
-            .border(
-                BorderStroke(1.dp, Color.Gray),
-                shape = RoundedCornerShape(20)
-            )
-            .clickable { onClick.invoke() }
-            .shadow(1.dp, shape = RoundedCornerShape(20))
-            .background(Color.White)
-    ) {
-        Icon(
-            Icons.Filled.Menu,
-            contentDescription = stringResource(id = R.string.menu),
-            tint = Color.Gray,
-            modifier = Modifier.size(22.dp)
-        )
-    }
-
-}
-
-@ExperimentalMaterial3Api
-@Composable
-fun MenuContent(
-    isOpen: Boolean,
-    childFullName: String,
-    parentFullName: String,
-    onClose: () -> Unit
-) {
-    val scope = rememberCoroutineScope()
-    val state = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-
-    LaunchedEffect(isOpen) {
-        if (isOpen) state.show() 
-    }
-
-    val onCloseClick: () -> Unit = {
-        scope.launch {
-            state.hide()
-            delay(300) // to await close animation
-            onClose.invoke()
-        }
-    }
-
-    if (isOpen) {
-        val scrollState = rememberScrollState()
-        ModalBottomSheet(
-            sheetState = state,
-            onDismissRequest = onCloseClick,
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(24.dp)
-            ) {
-                MenuHeader(onCloseClick = onCloseClick)
-                Column(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .verticalScroll(scrollState),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                ) {
-
-                    MenuSection(R.string.profile) {
-                        MenuItem(
-                            smallText = stringResource(id = R.string.child_name),
-                            mediumText = childFullName,
-                            imageId = R.drawable.seedling,
-                        )
-                        MenuItem(
-                            smallText = stringResource(id = R.string.parent_name),
-                            mediumText = parentFullName,
-                            imageId = R.drawable.tree,
-                        )
-                    }
-                    MenuSection(R.string.configurations) {
-                        MenuItem(
-                            smallText = stringResource(id = R.string.access_control),
-                            mediumText = stringResource(id = R.string.security),
-                            isReverseTextOrder = true,
-                            imageId = R.drawable.shield,
-                        )
-                        MenuItem(
-                            mediumText = stringResource(id = R.string.faq),
-                            imageId = R.drawable.question_mark,
-                        )
-                        MenuItem(
-                            mediumText = stringResource(id = R.string.privacy_term),
-                            imageId = R.drawable.clipboard,
-                        )
-                        MenuItem(
-                            mediumText = stringResource(id = R.string.exit_app),
-                            backgroundColor = LBLightPink,
-                            paddingTop = 24.dp,
-                            hasArrowRight = false,
-                            imageId = R.drawable.exit_door,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun MenuHeader(onCloseClick: () -> Unit) {
-    Box(modifier = Modifier.fillMaxWidth()) {
-        Text(
-            text = stringResource(id = R.string.menu),
-            style = MaterialTheme.typography.titleLarge,
-            fontWeight = FontWeight.Medium,
-            color = LBShadowGray,
-            modifier = Modifier.align(Alignment.Center)
-        )
-        Icon(
-            Icons.Filled.Close,
-            contentDescription = stringResource(id = R.string.close_menu),
-            tint = LBMediumGray,
-            modifier = Modifier
-                .size(30.dp)
-                .align(Alignment.CenterEnd)
-                .clickable { onCloseClick.invoke() },
-        )
-    }
-}
-
-@Composable
-fun MenuSection(@StringRes titleId: Int, content: @Composable () -> Unit) {
-    Column(modifier = Modifier.padding(top = 24.dp, bottom = 0.dp)) {
-        Text(
-            text = stringResource(id = titleId),
-            style = MaterialTheme.typography.titleLarge,
-            fontWeight = FontWeight.Medium,
-            color = LBMediumGray,
-        )
-        Column {
-            content.invoke()
-        }
-    }
-}
-
-@Composable
-fun MenuItem(
-    smallText: String = "",
-    mediumText: String,
-    @DrawableRes imageId: Int,
-    isReverseTextOrder: Boolean = false,
-    backgroundColor: Color = LBSoftBlue,
-    paddingTop: Dp = 16.dp,
-    hasArrowRight: Boolean = true,
-) {
-    Spacer(modifier = Modifier.padding(top = paddingTop))
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(90.dp)
-            .background(backgroundColor, shape = RoundedCornerShape(12.dp))
-            .border(1.dp, LBLightGray, shape = RoundedCornerShape(12.dp))
-            .padding(10.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            MenuImage(imageId, smallText)
-            Column(
-                Modifier
-                    .fillMaxWidth(0.8f)
-                    .padding(horizontal = 20.dp),
-                verticalArrangement = Arrangement.Center
-            ) {
-                if (isReverseTextOrder) {
-                    MenuMediumText(mediumText)
-                    MenuSmallText(smallText)
-                } else {
-                    MenuSmallText(smallText)
-                    MenuMediumText(mediumText)
-                }
-            }
-
-        }
-        if (hasArrowRight) {
-            Icon(
-                Icons.AutoMirrored.Filled.KeyboardArrowRight,
-                contentDescription = "",
-                modifier = Modifier.size(30.dp),
-                tint = LBSilverGray,
-            )
-        }
-    }
-}
-
-@Composable
-fun MenuImage(@DrawableRes imageId: Int, contentDescription: String) {
-    Surface(
-        modifier = Modifier
-            .size(45.dp)
-            .clip(CircleShape)
-            .background(Color.White),
-        shape = CircleShape
-    ) {
-        Box(contentAlignment = Alignment.Center) {
-            Icon(
-                painter = painterResource(id = imageId),
-                contentDescription = contentDescription,
-                tint = Color.Unspecified,
-                modifier = Modifier.size(24.dp)
-            )
-        }
-    }
-}
-
-
-@Composable
-fun MenuSmallText(text: String, modifier: Modifier = Modifier) {
-    if (text.isNotEmpty()) {
-        Text(
-            text = text,
-            color = Color.Gray,
-            style = MaterialTheme.typography.labelLarge,
-            modifier = modifier,
-        )
-    }
-}
-
-@Composable
-fun MenuMediumText(text: String, modifier: Modifier = Modifier) {
-    Text(
-        text = text,
-        style = MaterialTheme.typography.titleLarge,
-        fontWeight = FontWeight.Medium,
-        color = LBShadowGray,
-        modifier = modifier
-    )
-}
-
-@Preview
-@Composable
-fun MenuItemPreview() {
-    MenuItem("Nome da criança", "Maria da Silva", imageId = R.drawable.seedling)
-
-}
