+++ feature/dashboard/src/main/res/values-pt-rBR/strings.xml
@@ -2,7 +2,7 @@
 <resources>
     <string name="menu">Menu</string>
     <string name="logo_home">Logo LoryBlu</string>
-    <string name="hello">Olá, %s</string>
+    <string name="hello">Olá, %s!</string>
     <string name="close_menu">Fechar Menu</string>
     <string name="profile">Profile</string>
     <string name="child_name">Child\'s Name</string>
@@ -11,7 +11,7 @@
     <string name="security">Security</string>
     <string name="access_control">Control access and manage the app\'s functionalities</string>
     <string name="faq">Frequently Asked Questions</string>
-    <string name="privacy_term">Privacy Policy</string>
+    <string name="term_and_privacy">Privacy Policy</string>
     <string name="exit_app">Exit App</string>
 </resources>
